requirements.txt
python-dotenv>=1.0.0
openai>=1.2.0
asyncio>=3.4.3
xrpl-py>=2.0.0
postfiat>=0.1.0 